put init code into HardwareSimulator Class
